backend/pyproject.toml
[project]
name = "backend-django-elitemotors"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "GuilhermeHalter", email = "halter.guilhermenunes@gmail.com"},
]
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"
